refactor(modifier): make player health modifier more general
