+++ src/main/java/world/anhgelus/architectsland/difficultydeathscaler/difficulty/modifier/PlayerHealthModifier.java
@@ -1,4 +1,4 @@
-package world.anhgelus.architectsland.difficultydeathscaler.difficulty.player;
+package world.anhgelus.architectsland.difficultydeathscaler.difficulty.modifier;
 
 import net.minecraft.entity.attribute.EntityAttribute;
 import net.minecraft.entity.attribute.EntityAttributeModifier;
@@ -8,13 +8,13 @@ import net.minecraft.server.network.ServerPlayerEntity;
 import net.minecraft.util.Identifier;
 import world.anhgelus.architectsland.difficultydeathscaler.difficulty.DifficultyManager;
 
-public class PlayerHealthModifier extends DifficultyManager.Modifier {
-    public static final Identifier ID = Identifier.of("death_difficulty_health_modifier");
+public abstract class PlayerHealthModifier extends DifficultyManager.Modifier {
     public static final RegistryEntry<EntityAttribute> ATTRIBUTE = EntityAttributes.GENERIC_MAX_HEALTH;
     public static final EntityAttributeModifier.Operation OPERATION = EntityAttributeModifier.Operation.ADD_VALUE;
+    protected static Identifier IDENTIFIER;
 
-    public PlayerHealthModifier() {
-        super(ID, ATTRIBUTE, OPERATION);
+    public PlayerHealthModifier(Identifier id) {
+        super(id, ATTRIBUTE, OPERATION);
     }
 
     @Override
@@ -23,6 +23,6 @@ public class PlayerHealthModifier extends DifficultyManager.Modifier {
     }
 
     public static void apply(ServerPlayerEntity player, double value) {
-        apply(ID, ATTRIBUTE, OPERATION, player, value);
+        apply(IDENTIFIER, ATTRIBUTE, OPERATION, player, value);
     }
 }
