+++ src/main/java/world/anhgelus/architectsland/difficultydeathscaler/difficulty/player/PlayerDifficultyManager.java
@@ -3,12 +3,14 @@ package world.anhgelus.architectsland.difficultydeathscaler.difficulty.player;
 import net.minecraft.server.MinecraftServer;
 import net.minecraft.server.network.ServerPlayerEntity;
 import net.minecraft.text.Text;
+import net.minecraft.util.Identifier;
 import net.minecraft.world.Difficulty;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 import world.anhgelus.architectsland.difficultydeathscaler.DifficultyDeathScaler;
 import world.anhgelus.architectsland.difficultydeathscaler.difficulty.DifficultyManager;
 import world.anhgelus.architectsland.difficultydeathscaler.difficulty.StateSaver;
+import world.anhgelus.architectsland.difficultydeathscaler.difficulty.modifier.PlayerHealthModifier;
 
 import java.util.List;
 
@@ -19,26 +21,26 @@ public class PlayerDifficultyManager extends DifficultyManager {
 
     public static final StepPair[] STEPS = new StepPair[]{
             new StepPair(0, (server, gamerules, updater) -> {
-                updater.getModifier(PlayerHealthModifier.class).update(0);
+                updater.getModifier(HealthModifier.class).update(0);
             }),
             new StepPair(3, (server, gamerules, updater) -> {
-                updater.getModifier(PlayerHealthModifier.class).update(-2);
+                updater.getModifier(HealthModifier.class).update(-2);
             }),
             new StepPair(5, (server, gamerules, updater) -> {
-                updater.getModifier(PlayerHealthModifier.class).update(-4);
+                updater.getModifier(HealthModifier.class).update(-4);
             }),
             new StepPair(7, (server, gamerules, updater) -> {
-                updater.getModifier(PlayerHealthModifier.class).update(-6);
+                updater.getModifier(HealthModifier.class).update(-6);
             }),
             new StepPair(10, (server, gamerules, updater) -> {
-                updater.getModifier(PlayerHealthModifier.class).update(-8);
+                updater.getModifier(HealthModifier.class).update(-8);
             }),
             new StepPair(15, (server, gamerules, updater) -> {
-                updater.getModifier(PlayerHealthModifier.class).update(-10);
+                updater.getModifier(HealthModifier.class).update(-10);
             }),
     };
 
-    protected int playerHealthModifierValue = 0;
+    protected int healthModifierValue = 0;
 
     public PlayerDifficultyManager(MinecraftServer server, ServerPlayerEntity player) {
         super(server, STEPS, SECONDS_BEFORE_DECREASED);
@@ -50,6 +52,18 @@ public class PlayerDifficultyManager extends DifficultyManager {
         delayFirstTask(state.timeBeforeReduce);
 
         updateModifiersValue(modifiers(numberOfDeath));
+    }
+
+    public static class HealthModifier extends PlayerHealthModifier {
+        public static final Identifier ID = Identifier.of(PREFIX + "player_health_modifier");
+
+        static {
+            IDENTIFIER = ID;
+        }
+
+        public HealthModifier() {
+            super(ID);
+        }
     }
 
     @Override
@@ -67,14 +81,14 @@ public class PlayerDifficultyManager extends DifficultyManager {
 
     private void updateModifiersValue(List<Modifier> modifiers) {
         modifiers.forEach(m -> {
-            if (m instanceof final PlayerHealthModifier phm) playerHealthModifierValue = (int) phm.getValue();
+            if (m instanceof final HealthModifier phm) healthModifierValue = (int) phm.getValue();
             m.apply(player);
         });
     }
 
     @Override
     protected @NotNull String generateDifficultyUpdate(UpdateType updateType, @Nullable Difficulty difficulty) {
-        final var heartAmount = (20 + playerHealthModifierValue) / 2;
+        final var heartAmount = (20 + healthModifierValue) / 2;
 
         final var sb = new StringBuilder();
         sb.append(generateHeaderUpdate(updateType));
@@ -108,6 +122,6 @@ public class PlayerDifficultyManager extends DifficultyManager {
     }
 
     public void applyModifiers() {
-        PlayerHealthModifier.apply(player, playerHealthModifierValue);
+        HealthModifier.apply(player, healthModifierValue);
     }
 }
